+++ class/Config.php
@@ -5,11 +5,11 @@
  * Simple password manager written in PHP with Bootstrap and PDO database connections
  *
  *  File name: Config.php
- *  Last Modified: 10.01.23 г., 20:05 ч.
+ *  Last Modified: 19.01.23 г., 22:46 ч.
  *
  *  @link          https://blacktiehost.com
  *  @since         1.0.0
- *  @version       2.4.0
+ *  @version       3.0.0
  *  @author        Milen Karaganski <milen@blacktiehost.com>
  *
  *  @license       GPL-3.0+
@@ -18,107 +18,68 @@
  *
  */
 
-/**
- * \file        class/Config.php
- * \ingroup     Password Manager
- * \brief       This file is a config file for config class
- */
-
 declare(strict_types=1);
 
 namespace PasswordManager;
 
-use PasswordManagerCore\Triggers;
 use PDO;
 
 /**
  * Class for config
  */
-class Config
+abstract class Config
 {
     /**
-     * @var string Database host
+     * @var array Array of config data
      */
-    public string $host;
-    /**
-     * @var int Database port
-     */
-    public int $port;
+    private array $configData;
 
     /**
-     * @var string Database name
+     * @var PDO Database connection
      */
-    public string $dbname;
-
-    /**
-     * @var string Database tables prefix
-     */
-    public string $dbprefix;
-
-    /**
-     * @var string Database username
-     */
-    public string $dbuser;
-
-    /**
-     * @var string Database user password
-     */
-    public string $dbpass;
-
-    /**
-     * @var string Database character set
-     */
-    public string $db_character_set;
-
-    /**
-     * @var string Database collation
-     */
-    public string $db_collation;
-    /**
-     * @var string
-     */
-    public string $main_url_root;
-    /**
-     * @var string
-     */
-    public string $main_app_root;
-    /**
-     * @var string
-     */
-    public string $main_application_title;
+    private PDO $conn;
 
     public function __construct()
     {
 
+        // include the config file
         if (file_exists('../conf/conf.php')) {
             include_once '../conf/conf.php';
         } elseif (file_exists('../../conf/conf.php')) {
             include_once '../../conf/conf.php';
         }
 
-        //Define database variables from conf file
-        $this->host = $db_host;
-        $this->port = (int)$db_port;
-        $this->dbname = $db_name;
-        $this->dbprefix = $db_prefix;
-        $this->dbuser = $db_user;
-        $this->dbpass = $db_pass;
-        $this->db_character_set = $main_db_character_set;
-        $this->db_collation = $main_db_collation;
-        $this->main_url_root = $main_url_root;
-        $this->main_app_root = $main_app_root;
-        $this->main_application_title = $main_application_title;
+        // Define the config data
+        $this->configData = [
+            'host'                   => $db_host,
+            'port'                   => (int)$db_port,
+            'dbname'                 => $db_name,
+            'dbprefix'               => $db_prefix,
+            'dbuser'                 => $db_user,
+            'dbpass'                 => $db_pass,
+            'db_character_set'       => $main_db_character_set,
+            'db_collation'           => $main_db_collation,
+            'main_url_root'          => $main_url_root,
+            'main_app_root'          => $main_app_root,
+            'main_application_title' => $main_application_title,
+        ];
 
         //Connect to database and initialize global options and constants
         // For code consistency, all constants must be of type PM_*
         // with value 0 or 1 (false/true)
-        $conn = new PDO("mysql:host=$db_host;dbname=$db_name;port=$db_port", $db_user, $db_pass);
+        $this->conn = new PDO(
+            'mysql:host=' . $this->configData['host'] . ';
+                dbname=' . $this->configData['dbname'] . ';
+                port=' . $this->configData['port'],
+            $this->configData['dbuser'],
+            $this->configData['dbpass']
+        );
 
-        $sql = 'SELECT name, value from ' . $db_prefix . 'options';
-        $query = $conn->prepare($sql);
+        $sql = 'SELECT name, value from ' . $this->configData['dbprefix'] . 'options';
+        $query = $this->conn->prepare($sql);
 
-        if (!$conn->inTransaction()) {
-            $conn->beginTransaction();
+        if (!$this->conn->inTransaction()) {
+            $this->conn->beginTransaction();
         }
 
         $query->execute();
@@ -136,6 +97,20 @@ class Config
             }
         }
 
+        define('PM_MAIN_URL_ROOT', $this->configData['main_url_root']);
+        define('PM_MAIN_APP_ROOT', $this->configData['main_app_root']);
+        define('PM_MAIN_APPLICATION_TITLE', $this->configData['main_application_title']);
+        define('PM_MAIN_DB_PREFIX', $this->configData['dbprefix']);
+
         return $this;
+    }
+
+    /**
+     * @return array
+     */
+    public function getConfigData(): array
+    {
+
+        return $this->configData ?? [];
     }
 }