+++ class/Domains.php
@@ -5,11 +5,11 @@
  * Simple password manager written in PHP with Bootstrap and PDO database connections
  *
  *  File name: Domains.php
- *  Last Modified: 10.01.23 г., 20:17 ч.
+ *  Last Modified: 20.01.23 г., 8:52 ч.
  *
  *  @link          https://blacktiehost.com
  *  @since         1.0.0
- *  @version       2.4.0
+ *  @version       3.0.0
  *  @author        Milen Karaganski <milen@blacktiehost.com>
  *
  *  @license       GPL-3.0+
@@ -80,22 +80,20 @@ class Domains
     public string $child_table_element = 'records';
 
     /**
-     * @var array Array of fields to fetch from database
-     */
-    public array $array_of_fields = ['label', 'website', 'ftp', 'data_base', 'fk_user'];
-
-    /**
      * @var PassManDb Database handler
      */
     private PassManDb $db;
 
     /**
+     *    Constructor of the class
+     *
      * @param PassManDb $db Database handler
      */
     public function __construct(PassManDb $db)
     {
 
         $this->db = $db;
+        $this->db->error = '';
     }
 
     /**
@@ -110,18 +108,21 @@ class Domains
         if (empty(DISABLE_SYSLOG)) {
             pm_syslog(__METHOD__ . ' called from ' . get_class($this), PM_LOG_INFO);
         }
-        $array = [];
-        foreach ($this->array_of_fields as $val) {
-            if (!empty($this->$val)) {
-                $array[$val] = $this->$val;
-            }
-        }
 
-        $result = $this->db->create($array, $this->table_element);
+        $data = [
+            'label'     => $this->label,
+            'website'   => $this->website,
+            'ftp'       => $this->ftp,
+            'data_base' => $this->data_base,
+            'fk_user'   => $this->fk_user,
+        ];
+        $result = $this->db->create($this->table_element, $data);
 
         if ($result > 0) {
             return 1;
         } else {
+            $_SESSION['PM_ERROR'] = $this->db->error;
+            $this->error = $this->db->error;
             return -1;
         }
     }
@@ -129,27 +130,30 @@ class Domains
     /**
      * Update record in database
      *
-     * @param array $fields Array fo fields to update
-     *
      * @return int 1 if OK, <0 if KO
      * @throws PDOException|Exception
      */
-    public function update(array $fields)
+    public function update()
     {
 
         if (empty(DISABLE_SYSLOG)) {
             pm_syslog(__METHOD__ . ' called from ' . get_class($this), PM_LOG_INFO);
         }
-        $array_to_update = [];
-        foreach ($fields as $field) {
-            $array_to_update[$field] = $this->$field;
+
+        $data = [
+            'label' => $this->label,
+        ];
+        if (!empty($this->password)) {
+            $data['password'] = password_hash($this->password, PASSWORD_DEFAULT);
         }
 
-        $result = $this->db->update($array_to_update, $this->table_element, $this->id);
+        $result = $this->db->update($this->table_element, $data, "rowid = $this->id");
 
         if ($result > 0) {
             return 1;
         } else {
+            $_SESSION['PM_ERROR'] = $this->db->error;
+            $this->error = $this->db->error;
             return -1;
         }
     }
@@ -158,7 +162,7 @@ class Domains
      * Delete record from database
      *
      * @return int 1 if OK, <0 if KO
-     * @throws PDOException|Exception
+     * @throws Exception
      */
     public function delete()
     {
@@ -166,116 +170,113 @@ class Domains
         if (empty(DISABLE_SYSLOG)) {
             pm_syslog(__METHOD__ . ' called from ' . get_class($this), PM_LOG_INFO);
         }
-        $res = $this->db->fetchAll([], $this->child_table_element, ['fk_domain' => $this->id], '');
 
-        foreach ($res as $child) {
-            $this->db->delete($this->child_table_element, (int)$child['id']);
+        $res = $this->db->fetchAll(
+            $this->child_table_element,
+            'fk_domain = :fk_domain',
+            [':fk_domain' => $this->id]
+        );
+
+        if ($res) {
+            foreach ($res as $child) {
+                $result = $this->db->delete($this->child_table_element, 'rowid = :id', [':id' => (int)$child['rowid']]);
+                if ($result < 1) {
+                    $_SESSION['PM_ERROR'] = $this->db->error;
+                    $this->error = $this->db->error;
+                    return -1;
+                }
+            }
         }
 
-        $result = $this->db->delete($this->table_element, $this->id);
+        $delete = $this->db->delete($this->table_element, 'rowid = :id', [':id' => $this->id]);
 
-        if ($result > 0) {
+        if ($delete > 0) {
             return 1;
         } else {
-            return -1;
+            $_SESSION['PM_ERROR'] = $this->db->error;
+            $this->error = $this->db->error;
+            return -2;
         }
     }
 
     /**
-     * Fetch all records from database into array
+     * Fetch all records from database
      *
-     * @param array  $filter          Array of filters. Example array:('field' => 'value'). If key is customsql,
-     *                                it should be an array also like ('customsql' => array('field' = > 'value'))
-     * @param string $filter_mode     Filter mode AND or OR. Default is AND
-     * @param string $sortfield       Sort field
-     * @param string $sortorder       Sort order
-     * @param string $group           Group BY field name
-     * @param int    $limit           Limit
-     * @param int    $offset          Offset
+     * @param string $where  The WHERE clause of the query
+     * @param array  $params An array of parameters to bind to the query
      *
-     * @return int|array|false|string
-     * @throws PDOException|Exception
+     * @return array|int Array of objects or <0 on error
+     * @throws Exception
      */
-    public function fetchAll(
-        $filter = '',
-        string $filter_mode = 'AND',
-        string $sortfield = '',
-        string $sortorder = '',
-        string $group = '',
-        int $limit = 0,
-        int $offset = 0
-    ) {
+    public function fetchAll(string $where, array $params = [])
+    {
 
         if (empty(DISABLE_SYSLOG)) {
             pm_syslog(__METHOD__ . ' called from ' . get_class($this), PM_LOG_INFO);
         }
 
-        $result = $this->db->fetchAll(
-            $this->array_of_fields,
-            $this->table_element,
-            $filter,
-            $filter_mode,
-            $sortfield,
-            $sortorder,
-            $group,
-            $limit,
-            $offset
-        );
+        $result = $this->db->fetchAll($this->table_element, $where, $params);
+        $domains = [];
+        if ($result) {
+            foreach ($result as $row) {
+                $domain = new self($this->db);
+                $domain->id = (int)$row['rowid'];
+                $domain->label = $row['label'];
+                if ($row['website']) {
+                    $domain->website = (int)$row['website'];
+                }
+                if ($row['ftp']) {
+                    $domain->ftp = (int)$row['ftp'];
+                }
+                if ($row['data_base']) {
+                    $domain->data_base = (int)$row['data_base'];
+                }
+                $domain->fk_user = (int)$row['fk_user'];
+                $domains[] = $domain;
+            }
 
-        if ($result > 0) {
-            return $result;
+            return $domains;
         } else {
+            $_SESSION['PM_ERROR'] = $this->db->error;
+            $this->error = $this->db->error;
             return -1;
         }
     }
 
     /**
-     * Fetch single row from database
+     * Fetch a single record from database
      *
-     * @param int    $id              ID of the record to fetch
-     * @param array  $filter          Array of filters. Example array:('field' => 'value'). If key is customsql,
-     *                                it should be an array also like ('customsql' => array('field' = > 'value'))
-     * @param string $filter_mode     Filter mode AND or OR. Default is AND
-     * @param string $sortfield       Sort field
-     * @param string $sortorder       Sort order
-     * @param string $group           Group BY field name
-     * @param int    $limit           Limit
-     * @param int    $offset          Offset
+     * @param int $id The ID of the record to fetch
      *
-     * @return int
-     * @throws PDOException|Exception
+     * @return Domains|int Return Object or <0 on error
+     * @throws Exception
      */
-    public function fetch(
-        int $id,
-        $filter = '',
-        string $filter_mode = 'AND',
-        string $sortfield = '',
-        string $sortorder = '',
-        string $group = '',
-        int $limit = 0,
-        int $offset = 0
-    ) {
+    public function fetch(int $id)
+    {
 
         if (empty(DISABLE_SYSLOG)) {
             pm_syslog(__METHOD__ . ' called from ' . get_class($this), PM_LOG_INFO);
         }
 
-        $result = $this->db->fetch(
-            $id,
-            $this->array_of_fields,
-            $this->table_element,
-            $filter,
-            $filter_mode,
-            $sortfield,
-            $sortorder,
-            $group,
-            $limit,
-            $offset
-        );
+        $result = $this->db->fetch($this->table_element, 'rowid = :id', [':id' => $id]);
+        if ($result) {
+            $this->id = (int)$result['rowid'];
+            $this->label = $result['label'];
+            if ($result['website']) {
+                $this->website = (int)$result['website'];
+            }
+            if ($result['ftp']) {
+                $this->ftp = (int)$result['ftp'];
+            }
+            if ($result['data_base']) {
+                $this->data_base = (int)$result['data_base'];
+            }
+            $this->fk_user = (int)$result['fk_user'];
 
-        if ($result > 0) {
-            return $result;
+            return $this;
         } else {
+            $_SESSION['PM_ERROR'] = $this->db->error;
+            $this->error = $this->db->error;
             return -1;
         }
     }